# app/definition/swagger.yaml
swagger: "2.0"
info:
  version: desp - v1
  title: Despachalo.com
  description: API de servicios de Despachalo.com
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /desp/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /token:
    get:
      tags:
        - Token
      description: Lista los tipos de token (bolsa)
      # used as the method name of the controller
      operationId: findTokens
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: email
          in: query
          description: Correo del usuario
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Token
      description: Registrar Token
      # used as the method name of the controller
      operationId: addToken
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Token"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Token
      description: Editar Token
      # used as the method name of the controller
      operationId: updateToken
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Token"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Token
      description: Elimina un Token
      # used as the method name of the controller
      operationId: deleteToken
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Token"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Token:
    type: object
    required:
      - email
      - name
    properties:
      email:
        description: Correo del usuario Admin
        type: string
      name:
        description: Nombre
        type: string
      numQuote:
        description: Número de cotizaciones
        type: string
      rate:
        description: Costo del Token
        type: string
...
  DespachaloResponse:
    type: string
  ErrorResponse:
    type: string